starknet_transaction_prover: redact URL credentials at CLI-override log sites
The CLI-override `info!` calls in `ServiceConfig::from_args` echoed the
raw `rpc_node_url` and `blocking_check_url` strings, which can carry
credentials in the userinfo component (`https://user:secret@rpc...`).
Both URLs now go through `redact_url_host` so only `scheme://host[:port]`
reaches the log sink. The helper moves out of `main.rs` into
`server/log_redact.rs` so both the startup banner and the override logs
share one implementation.

Caught during a security review of the observability rollout.
